Sync

+++ acemeta/GitHub.py
@@ -1,5 +1,6 @@
 import requests as WEB
 import base64
+import os
 
 class Repository():
     """
@@ -10,6 +11,7 @@ class Repository():
         token (str): A PAT-Token that has access to the requested repository
 
     #### Methods:
+        exists(): Checks if a file exists in the repository
         upload(): Uploads a file to the repository
         download(): Downloads a file from the repository
     """
@@ -19,7 +21,29 @@ class Repository():
 
         self._url = f"https://api.github.com/repos/{self.repository}/contents"
 
-    def upload(self, file: str, directory: str, msg: str) -> None:
+    def exists(self, directory: str) -> bool:
+        """
+        Checks if a file already exists in the specified directory of the repository
+
+        #### Arguments:
+            directory (str): Path of the file whose existence is to be checked
+
+        #### Raises:
+            Exception: If the files existence couldn't be checked  
+        """
+        target = f"{self._url}/{directory}"
+        headers = {
+            'Authorization': f'token {self.token}',
+            'Accept': 'application/vnd.github.v3+json'
+        }
+
+        response = WEB.get(target, headers=headers)
+        
+        if response.status_code == 200: return True
+        elif response.status_code == 404: return False
+        else: response.raise_for_status()
+
+    def upload(self, file: str, directory: str, msg: str, overwrite: bool = False) -> None:
         """
         Uploads a file to the GitHub repository
 
@@ -27,45 +51,54 @@ class Repository():
             file (str): Path to the file to upload
             directory (str): Path of the file that is to be created
             msg (str): An optional upload information displayed as the commit summary
+            overwrite (bool): Whether the function is allowed to overwrite the file if it already exists
 
         #### Raises:
             FileExistsError: If a file with the same content already exists in the requested directory
+                Carries a status_code attribute containing the HTTP status code
             Exception: For other errors encountered during download
+                Carries a response.status_code attribute containing the HTTP status code
         """
 
-        target = f"{self._url}/{directory}"
+        if Repository.exists(directory) and overwrite == False:
+            raise FileExistsError(f"File '{self.repository}/{directory}' already exists.")
+        else: 
         
-        with open(file, 'rb') as f:
-            content = f.read()
-            content_base64 = base64.b64encode(content).decode("utf-8")
+            target = f"{self._url}/{directory}"
+            
+            with open(file, 'rb') as f:
+                content = f.read()
+                content_base64 = base64.b64encode(content).decode("utf-8")
 
-        headers = {
-            'Authorization': f'token {self.token}',
-            'Content-Type': 'application/json'
-        }
-        data = {
-        'message': msg,
-        'content': content_base64
-        }
-    
-        response = WEB.put(target, json=data, headers=headers)
+            headers = {
+                'Authorization': f'token {self.token}',
+                'Content-Type': 'application/json'
+            }
+            data = {
+            'message': msg,
+            'content': content_base64
+            }
+        
+            response = WEB.put(target, json=data, headers=headers)
 
-        if response.status_code == 422:
-            raise FileExistsError(f"File '{self.repository}/{directory}' already exists with the exact same content")
-        response.raise_for_status()
+            if response.status_code == 422:
+                raise FileExistsError(f"File '{self.repository}/{directory}' already exists with the exact same content", status_code=response.status_code)
+            response.raise_for_status()
 
 
-    def download(self, file: str, destination: str) -> None:
+    def download(self, file: str, destination: str, overwrite: bool = False) -> None:
         """
         Downloads a file from the GitHub repository.
 
         #### Arguments:
             file (str): Path to the file in the GitHub repository
             destination (str): Local path where the file should be saved
+            overwrite (bool): Whether the function is allowed to overwrite the file if it already exists
 
         #### Raises:
             FileNotFoundError: If the file is not found in the repository
-            Exception: For other errors encountered during download
+            FileExistsError: If a file is to be overwritten in the course of the download while the function is not allowed to do so
+            Exception: For other errors encountered during download 
         """
 
         target = f"{self._url}/contents/{file}"
@@ -84,5 +117,8 @@ class Repository():
         content_base64 = response.content
         content = base64.b64decode(content_base64)
         
-        with open(destination, 'wb') as f:
-            f.write(content)
+        if os.path.exists(destination) and overwrite == False:
+            raise FileExistsError(f"File '{destination}' already exists and mustn't be overwritten")
+        else:
+            with open(destination, 'wb') as f:
+                f.write(content)
